requirements.txt
Flask
boto3
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
gunicorn